Removed old reference to nav.css and added if around django's context
There was a reference to nav.css, which has been deleted, that was
causing trouble for testing (that error would somehow stop the test
server from getting torn down completely, leading to "Address already
in use" errors on any other tests that used StaticLiveServerTestCase).

If the quiz context from Django was not defined, it would replace the
template tag with nothing, leading to syntactically incorrect statements
like "context = { quiz: };". Surrounded the assignment in an if
statment, so it's only made if the context is defined. This was
potentially also contributing to the "Address already in use" issues.