Removed the ChoiceResponse model and cleaned up some model tests
Removed the ChoiceResponse model, because it's more direct to link
directly to Choice from CheckboxResponse. Also created a new fixture
to do testing with. Had some issues with migrating, unfortunately, but
doing it in two steps (first removing the ChoiceResponse model, then
adding a M2M field to linke CheckboxResponse to Choice) seemed to work
(I think because the 'choices' field in CheckboxResponse was going
directly from reverse-FK to M2M without being renamed).

This also prompted a simplification of the ModelDeleteTests and a
removal of some unnecessary tests of reverse-FK links.